Move EventRating to own class
